refactor: consolidate duplicated test fixtures into conftest files
- Create tests/clustering/conftest.py with shared well_separated_data
  fixture (was duplicated identically in test_density, test_auto,
  test_spectral)
- Create tests/imaging/conftest.py with shared sample_data fixture
  (was duplicated identically in test_angular and test_recurrence)
- Create tests/bayesian/conftest.py with shared local_level_matrices
  helper (was defined in test_kalman, will be reused by future
  Bayesian tests)
- Remove duplicate definitions from individual test files

Closes #147
