+++ tests/bayesian/test_kalman.py
@@ -4,19 +4,7 @@ import pytest
 
 from polars_ts.bayesian.kalman import KalmanFilter, KalmanResult, kalman_filter
 
-# --- Local-level model helpers ---
-# State: x_t (scalar level)
-# Observation: y_t = x_t + v_t
-# Transition: x_t = x_{t-1} + w_t
-
-
-def _local_level_matrices(q: float = 1.0, r: float = 1.0):
-    """Return system matrices for a local-level (random walk + noise) model."""
-    F = np.array([[1.0]])
-    H = np.array([[1.0]])
-    Q = np.array([[q]])
-    R = np.array([[r]])
-    return F, H, Q, R
+from .conftest import local_level_matrices as _local_level_matrices
 
 
 @pytest.fixture
