+++ tests/clustering/test_auto.py
@@ -9,26 +9,6 @@ _has_scipy = importlib.util.find_spec("scipy") is not None
 _has_sklearn = importlib.util.find_spec("sklearn") is not None
 
 
-@pytest.fixture
-def well_separated_data():
-    """Six series in two well-separated groups (ascending vs descending)."""
-    ascending = [1.0, 2.0, 3.0, 4.0]
-    descending = [4.0, 3.0, 2.0, 1.0]
-    return pl.DataFrame(
-        {
-            "unique_id": (["A1"] * 4 + ["A2"] * 4 + ["A3"] * 4 + ["B1"] * 4 + ["B2"] * 4 + ["B3"] * 4),
-            "y": (
-                ascending
-                + [1.0, 2.1, 3.0, 4.1]
-                + [1.0, 1.9, 3.1, 4.0]
-                + descending
-                + [4.1, 3.0, 2.0, 0.9]
-                + [3.9, 3.1, 1.9, 1.0]
-            ),
-        }
-    )
-
-
 class TestAutoClusterResult:
     def test_returns_auto_cluster_result(self, well_separated_data):
         result = auto_cluster(well_separated_data, methods=["kmedoids"], distances=["sbd"], k_range=range(2, 4))
